pyproject.toml
[project]
name = "rare-bird-discord-api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.92.0",
    "discord-py>=2.0",
    "loguru>=0.7.3",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "ruff>=0.15.9",
    "pytest>=8.3",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.24",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=apps --cov=core --cov-report=term-missing --cov-fail-under=85"

[tool.coverage.run]
branch = true
source = ["apps", "core"]
omit = ["tests/*", "main.py"]

[tool.coverage.report]
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
